fixed user controller

+++ app/Http/Controllers/UserController.php
@@ -17,6 +17,8 @@ use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
+    protected $redirectTo = '/home';
+
     /**
      * Create a new controller instance.
      *
@@ -35,46 +37,51 @@ class UserController extends Controller
         $user = User::where('email',$email)->first();
 
         if(!is_null($user)){
+
             $bio = UserBiography::where('user_id', $user->id )->first();
+
             $posts = Post::with('comments.author')->where('user_id', $user->id)->orderBy('created_at', 'desc')->get();
+
             return view('user.profile',compact('bio','user','posts'));
 
         }
+        
         return redirect()->route('home');
+
     }
 
     public function usernameEdit($id)
     {
-        // return $id;
+
         $user = User::where('id', $id)->first();
-        if(is_null( $user ) ){
-            return redirect()->route('home');
-        }
 
         //checks if the Authenticated user is able to see the edit comment
-        if( Auth::user()->id === $user->id ){
-            // return ['status'=>'true'];
+        if( !is_null( $user ) && Auth::user()->id === $user->id ){
+
             return view('forms.usernameEditForm',compact('user', 'id' ));
         }
-        else{
-            // return ['status'=>'false'];
-            return redirect()->route('home');
-        }
+
+        return redirect()->route('home');
+
 
     }
 
     public function usernameUpdate(Request $request)
     {
-        // return $request;
-        // return  Auth::user();
         if( Auth::user()->id === intval($request->user_id) ){
+
             $user = User::where('id', $request->user_id)
+
             ->update(['username' => $request->username]);
+
             return redirect()->route( 'profile', Auth::user()->email );
+
         }
+
         else{
-            // return ['status' => 'false'];
+
             return redirect()->route('home');
+
         }
 
     }
@@ -82,29 +89,41 @@ class UserController extends Controller
     public function delete($user_id)
     {
         $user = User::where('id',$user_id)->first();
-        // return $user;
-        Mail::to($user->email)->send(new UserDeletionEmail($user));
 
-        $user = User::where('id',$user_id)->delete();
+        if( !is_null( $user ) && (Auth::user()->id === $user->id || Auth::user()->admin === 1)  ){
+
+            Mail::to($user->email)->send(new UserDeletionEmail($user));
+
+            $user = User::where('id',$user_id)->delete();
+
+        }
         return redirect()->route('home');
 
     }
 
     public function makeAdmin($id)
     {
-
         $user = User::where('id',$id)->first();
-        $userAfter = User::where('id',$id)->update(['admin'=>1]);
-        // return $userAfter;
+
+        if( !is_null( $user ) && ( Auth::user()->admin === 1)  ){
+
+            $userAfter = User::where('id',$id)->update(['admin'=>1]);
+
+        }
+
         return back()->withInput();
     }
 
     public function removeAdmin($id)
     {
-
         $user = User::where('id',$id)->first();
-        $userAfter = User::where('id',$id)->update(['admin'=>0]);
-        // return $userAfter;
+
+        if( !is_null( $user ) && ( Auth::user()->admin === 1)  ){
+
+            $userAfter = User::where('id',$id)->update(['admin'=>0]);
+        }
+
         return back()->withInput();
+
     }
 }
